+++ src/OpenCqrs.EventSourcing.Store.EntityFrameworkCore/Extensions/DbContextExtensions/GetInMemoryAggregate.cs
@@ -1,5 +1,4 @@
-﻿using System.Reflection;
-using OpenCqrs.EventSourcing.Domain;
+﻿using OpenCqrs.EventSourcing.Domain;
 using OpenCqrs.EventSourcing.Store.EntityFrameworkCore.Entities;
 using OpenCqrs.Results;
 
@@ -217,6 +216,24 @@ public static partial class IDomainDbContextExtensions
             ? await domainDbContext.GetEventEntitiesUpToSequence(streamId, upToSequence.Value, aggregate.EventTypeFilter, cancellationToken)
             : await domainDbContext.GetEventEntities(streamId, aggregate.EventTypeFilter, cancellationToken);
 
+        if (eventEntities.Count == 0)
+        {
+            return aggregate;
+        }
+
+        aggregate.StreamId = streamId.Id;
+        aggregate.AggregateId = aggregateId.ToStoreId();
+        aggregate.LatestEventSequence = eventEntities.OrderBy(eventEntity => eventEntity.Sequence).Last().Sequence;
+        aggregate.Apply(eventEntities.Select(eventEntity => eventEntity.ToDomainEvent()));
+
+        return aggregate;
+    }
+
+    public static async Task<Result<TAggregate>> GetInMemoryAggregate<TAggregate>(this IDomainDbContext domainDbContext, IStreamId streamId, IAggregateId<TAggregate> aggregateId, DateTimeOffset upToDate, CancellationToken cancellationToken = default) where TAggregate : IAggregate, new()
+    {
+        var aggregate = new TAggregate();
+
+        var eventEntities = await domainDbContext.GetEventEntitiesUpToDate(streamId, upToDate, aggregate.EventTypeFilter, cancellationToken);
         if (eventEntities.Count == 0)
         {
             return aggregate;